pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ipro"
version = "1.5.0"
description = "CLI tool for responsive image processing"
requires-python = ">=3.8"
license = "MIT"
dependencies = [
    "Pillow>=10.0.0",
    "pillow-heif>=0.13.0",
]

[project.scripts]
ipro = "ipro:main"

[tool.setuptools]
py-modules = ["ipro"]
